protect against zero charge particles
Former-commit-id: 829f81f9a2cb40df38b732da70137d7b6d25fefb

+++ tracker/TRKFactory.hh
@@ -52,7 +52,7 @@ private:
   //TRKElement* createTransform3D(Element& element);
 
   /// particle properties
-  double charge;
+  int charge;
   /// beam momentum in GeV
   double momentum;
   double brho;
